채팅방 리스트, 채팅방 페이지 통합
